Add support for GUI modal dialogs on calls to confirm(...)
WARNING: This feature is dependent on proposed changes to NeoVim. A new
property "ext_dialogs" must be added for this to work.

Most editors support GUI dialog prompts for save confirmation. Currently
NeoVim handles this in the cmdline area. This change intercepts calls to
Vim's confirm(...) function, and displays a GUI modal dialog instead.