Refactor NeovimConnector Creation
On MacOs, the OS provides a special key sequence (Cmd + `) to switch
between windows owned by the same process. In order to properly support
multiple windows, we need to decouple the connection from App so that
the connection can be owned by the window and the proper clean up happens
when the window exits.

Refactor only, no behavior changes.